create route correctly
