+++ src/app/layout.js
@@ -1,0 +1,26 @@
+import { Inter } from "next/font/google";
+import "../styles/globals.css";
+import "../styles/tailwind.css";
+import Header from "../components/Header";
+import Footer from "../components/Footer";
+
+const inter = Inter({ subsets: ["latin"] });
+
+export const metadata = {
+  title: "Create Next App",
+  description: "Generated by create next app",
+};
+
+export default function RootLayout({ children }) {
+  return (
+    <html lang="en">
+      <body className={inter.className}>
+        <div className="flex flex-col h-screen bg-[#000000]">
+          <Header />
+          {children}
+          <Footer />
+        </div>
+      </body>
+    </html>
+  );
+}
