+++ src/components/Header.js
@@ -1,10 +1,7 @@
 import Link from "next/link";
 import styles from "../styles/Header.module.css";
-import { useRouter } from "next/router";
 
 export default function Header() {
-  const router = useRouter();
-
   return (
     <header className="bg-[#023047] p-4 flex flex justify-between">
       <div className="left-div">
@@ -16,42 +13,22 @@ export default function Header() {
         <nav className="right-div">
           <ul className="flex gap-5 p-0 m-0 list-none">
             <li className={styles.li}>
-              <Link
-                href="/"
-                className={`${styles.link} ${
-                  router.pathname === "/" ? styles.activeLink : ""
-                }`}
-              >
+              <Link href="/" className={styles.link}>
                 Home
               </Link>
             </li>
             <li className={styles.li}>
-              <Link
-                href="/about"
-                className={`${styles.link} ${
-                  router.pathname === "/about" ? styles.activeLink : ""
-                }`}
-              >
+              <Link href="/about" className={styles.link}>
                 About
               </Link>
             </li>
             <li className={styles.li}>
-              <Link
-                href="/projects"
-                className={`${styles.link} ${
-                  router.pathname === "/projects" ? styles.activeLink : ""
-                }`}
-              >
+              <Link href="/projects" className={styles.link}>
                 Projects
               </Link>
             </li>
             <li className={styles.li}>
-              <Link
-                href="/contact"
-                className={`${styles.link} ${
-                  router.pathname === "/contact" ? styles.activeLink : ""
-                }`}
-              >
+              <Link href="/contact" className={styles.link}>
                 Contact
               </Link>
             </li>
