+++ src/components/IconButton.js
@@ -1,3 +1,5 @@
+"use client";
+
 import { useRef, useState } from "react";
 
 export default function IconButton({ children, text, color, ...props }) {
